+++ apps/alpha-test-app/src/views/FormDemoView.vue
@@ -1,8 +1,7 @@
 <script lang="ts" setup>
 import { OnyxAppLayout, OnyxHeadline, OnyxPageLayout } from "sit-onyx";
 import { useI18n } from "vue-i18n";
-import LanguageSelection from "../components/LanguageSelection.vue";
-import FormDemo from "../components/form-demo/FormDemo.vue";
+import FormDemo, { type FormData } from "../components/form-demo/FormDemo.vue";
 
 const { t, locale } = useI18n();
 
@@ -10,18 +9,21 @@ const validFormData = {
   defaultInput: "No Validation",
   requiredInput: "Is filled",
   minlengthInput: "Is long enough",
-  maxInput: 42,
   typeInput: "john.doe@mail.schwarz",
   patternInput: "only lowercase or space",
-};
+  switch: true,
+  checkboxGroup: [2],
+  radioGroup: { id: 1, label: "Option 1" },
+} satisfies FormData;
+
 const invalidFormData = {
   defaultInput: "No Validation",
   requiredInput: "",
   minlengthInput: "",
-  maxInput: 9001,
   typeInput: "NotAmail",
   patternInput: "NO UPPERCASE ALLOWED",
-};
+  switch: false,
+} satisfies FormData;
 </script>
 
 <template>