+++ packages/sit-onyx/src/components/OnyxInput/OnyxInput.vue
@@ -1,9 +1,10 @@
 <script lang="ts" setup>
 import { computed } from "vue";
+import { useDensity } from "../../composables/density";
+import { useRequired } from "../../composables/required";
+import { useCustomValidity } from "../../composables/useCustomValidity";
 import OnyxLoadingIndicator from "../OnyxLoadingIndicator/OnyxLoadingIndicator.vue";
 import type { OnyxInputProps } from "./types";
-import { useRequired } from "../../composables/required";
-import { useDensity } from "../../composables/density";
 
 const props = withDefaults(defineProps<OnyxInputProps>(), {
   modelValue: "",
@@ -32,7 +33,13 @@ const emit = defineEmits<{
    * Emitted when the input is blurred.
    */
   blur: [];
+  /**
+   * Emitted when the validity state of the input changes.
+   */
+  validityChange: [validity: ValidityState];
 }>();
+
+const { vCustomValidity } = useCustomValidity({ props, emit });
 
 const { requiredMarkerClass, requiredTypeClass } = useRequired(props);
 const { densityClass } = useDensity(props);
@@ -77,6 +84,7 @@ const shouldShowCounter = computed(() => props.withCounter && props.maxlength);
       -->
         <input
           v-model="value"
+          v-custom-validity
           class="onyx-input__native"
           :placeholder="props.placeholder"
           :type="props.type"
@@ -146,15 +154,14 @@ const shouldShowCounter = computed(() => props.withCounter && props.maxlength);
   @include density.cozy {
     --onyx-input-padding-vertical: var(--onyx-spacing-sm);
   }
-}
 
-.onyx-input {
   @include layers.component() {
     --border-color: var(--onyx-color-base-neutral-300);
     --selection-color: var(--onyx-color-base-primary-200);
+    --outline-color: var(--onyx-color-base-primary-200);
 
     font-family: var(--onyx-font-family);
-    display: block;
+    display: flex;
     flex-direction: column;
     gap: var(--onyx-spacing-5xs);
 
@@ -183,17 +190,16 @@ const shouldShowCounter = computed(() => props.withCounter && props.maxlength);
 
       font-size: 1rem;
       line-height: $line-height;
-
       padding: var(--onyx-input-padding-vertical) var(--onyx-spacing-sm);
       height: calc($line-height + 2 * var(--onyx-input-padding-vertical));
 
       &:has(.onyx-input__native:read-write:hover) {
-        border-color: var(--onyx-color-base-primary-400);
+        --border-color: var(--onyx-color-base-primary-400);
       }
 
       &:has(.onyx-input__native:enabled:focus) {
         --border-color: var(--onyx-color-base-primary-500);
-        outline: var(--onyx-spacing-4xs) solid var(--onyx-color-base-primary-200);
+        outline: var(--onyx-spacing-4xs) solid var(--outline-color);
       }
 
       // :read-only is valid for readonly and disabled state so we put shared styles for both states here
@@ -217,6 +223,20 @@ const shouldShowCounter = computed(() => props.withCounter && props.maxlength);
       &:has(.onyx-input__native:read-write) {
         &:has(#{get-autofill-selectors(".onyx-input__native")}) {
           background-color: var(--onyx-color-base-warning-100);
+        }
+      }
+
+      &:has(.onyx-input__native:user-invalid) {
+        --border-color: var(--onyx-color-base-danger-500);
+        --outline-color: var(--onyx-color-base-danger-200);
+        --selection-color: var(--onyx-color-base-danger-200);
+
+        &:has(.onyx-input__native:enabled:focus) {
+          --border-color: var(--onyx-color-base-danger-500);
+        }
+
+        &:has(.onyx-input__native:enabled:hover) {
+          --border-color: var(--onyx-color-base-danger-400);
         }
       }
     }