+++ packages/sit-onyx/src/index.ts
@@ -41,6 +41,7 @@ export * from "./components/OnyxPageLayout/types";
 
 export * from "./components/OnyxRadioButton/types";
 export { default as OnyxRadioButtonGroup } from "./components/OnyxRadioButtonGroup/OnyxRadioButtonGroup.vue";
+export * from "./components/OnyxRadioButtonGroup/types";
 
 export { default as OnyxSelect } from "./components/OnyxSelect/OnyxSelect.vue";
 export * from "./components/OnyxSelect/types";
@@ -53,9 +54,8 @@ export * from "./components/OnyxSwitch/types";
 export { default as OnyxTooltip } from "./components/OnyxTooltip/OnyxTooltip.vue";
 export * from "./components/OnyxTooltip/types";
 
-export { default as TestInput } from "./components/TestInput/TestInput.vue";
-
 export * from "./composables/density";
+export * from "./composables/scrollEnd";
 export type { OnyxTranslations, ProvideI18nOptions } from "./i18n";
 export * from "./types";
 export { createOnyx } from "./utils/plugin";